+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 
 Full changelog for PHP Quill Renderer
 
+## v3.14.2 - 2018-09-11
+
+* New line only inserts no longer ignored by parser, updated a couple of tests because of logic update.
+* Updated README, added new credit.
+
 ## v3.14.1 - 2018-08-20
 
 * Multiple attributes ignored by the Markdown and the GitHub Markdown parsers, added two tests to catch regressions.